IPFS stuff (#13)
* merged ipfs app to src

* updated git ignore

* Delete ProjectSettings.json

Deleted meta folder
