--- ipfs/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "ipfs",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "ipfs-api": "^22.1.1",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-scripts": "1.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  }
-}
